# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alter Event Tracker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/organizations/": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "List all organizations",
        "description": "List all organizations. Only SITE_ADMIN can see all orgs.",
        "operationId": "list_organizations_organizations__get",
        "responses": {
          "200": {
            "description": "All organizations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgListResponse"
                },
                "examples": {
                  "orgs_list": {
                    "summary": "Multiple organizations",
                    "value": {
                      "total": 2,
                      "organizations": [
                        {
                          "id": "c3d4e5f6-a7b8-9012-cdef-123456789012",
                          "name": "Alter Buenos Aires",
                          "slug": "alter-bsas",
                          "created_at": "2024-03-15T20:00:00Z"
                        },
                        {
                          "id": "d4e5f6a7-b8c9-0123-defa-234567890123",
                          "name": "Club Vanta",
                          "slug": "club-vanta",
                          "created_at": "2024-03-15T20:00:00Z"
                        }
                      ]
                    }
                  },
                  "empty": {
                    "summary": "No organizations yet",
                    "value": {
                      "total": 0,
                      "organizations": []
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Site admin role required",
            "content": {
              "application/json": {
                "examples": {
                  "site_admin_required": {
                    "summary": "Site admin privileges needed",
                    "value": {
                      "detail": "Site admin privileges required."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Create a new organization",
        "description": "Create a new organization. Only SITE_ADMIN can do this.",
        "operationId": "create_organization_organizations__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgCreate"
              },
              "examples": {
                "new_org": {
                  "summary": "Create a new organization",
                  "value": {
                    "name": "Alter Buenos Aires",
                    "slug": "alter-bsas"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Organization created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgPublic"
                },
                "examples": {
                  "created": {
                    "summary": "Organization created successfully",
                    "value": {
                      "id": "c3d4e5f6-a7b8-9012-cdef-123456789012",
                      "name": "Alter Buenos Aires",
                      "slug": "alter-bsas",
                      "created_at": "2024-03-15T20:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Name or slug already taken",
            "content": {
              "application/json": {
                "examples": {
                  "duplicate_name": {
                    "summary": "Name already taken",
                    "value": {
                      "detail": "Cannot create organization: name 'Alter Buenos Aires' is already taken."
                    }
                  },
                  "duplicate_slug": {
                    "summary": "Slug already taken",
                    "value": {
                      "detail": "Cannot create organization: slug 'alter-bsas' is already taken."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Site admin role required",
            "content": {
              "application/json": {
                "examples": {
                  "site_admin_required": {
                    "summary": "Site admin privileges needed",
                    "value": {
                      "detail": "Site admin privileges required."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/organizations/{org_id}": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Get a single organization",
        "description": "Get an organization. Available to members and SITE_ADMIN.",
        "operationId": "get_organization_organizations__org_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgPublic"
                },
                "examples": {
                  "org": {
                    "summary": "Single organization",
                    "value": {
                      "id": "c3d4e5f6-a7b8-9012-cdef-123456789012",
                      "name": "Alter Buenos Aires",
                      "slug": "alter-bsas",
                      "created_at": "2024-03-15T20:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Account not approved or disabled",
            "content": {
              "application/json": {
                "examples": {
                  "pending_approval": {
                    "summary": "Account pending approval",
                    "value": {
                      "detail": "Your account is pending admin approval. Please try again later."
                    }
                  },
                  "account_disabled": {
                    "summary": "Account disabled",
                    "value": {
                      "detail": "Your account has been disabled."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "examples": {
                  "org_not_found": {
                    "summary": "Organization UUID does not exist",
                    "value": {
                      "detail": "Organization c3d4e5f6-a7b8-9012-cdef-123456789012 not found."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{org_id}/members": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "List members of an organization",
        "description": "List all members of an organization with their org roles.",
        "operationId": "list_org_members_organizations__org_id__members_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members of this organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgMemberListResponse"
                },
                "examples": {
                  "members": {
                    "summary": "Mix of staff and admin members",
                    "value": {
                      "total": 2,
                      "members": [
                        {
                          "user_id": 1,
                          "org_id": "c3d4e5f6-a7b8-9012-cdef-123456789012",
                          "role": "ADMIN"
                        },
                        {
                          "user_id": 2,
                          "org_id": "c3d4e5f6-a7b8-9012-cdef-123456789012",
                          "role": "STAFF"
                        }
                      ]
                    }
                  },
                  "empty": {
                    "summary": "No members yet",
                    "value": {
                      "total": 0,
                      "members": []
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Site admin role required",
            "content": {
              "application/json": {
                "examples": {
                  "site_admin_required": {
                    "summary": "Site admin privileges needed",
                    "value": {
                      "detail": "Site admin privileges required."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "examples": {
                  "org_not_found": {
                    "summary": "Organization UUID does not exist",
                    "value": {
                      "detail": "Organization c3d4e5f6-a7b8-9012-cdef-123456789012 not found."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrgCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 2,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "title": "OrgCreate",
        "description": "Request body to create a new organization."
      },
      "OrgListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "organizations": {
            "items": {
              "$ref": "#/components/schemas/OrgPublic"
            },
            "type": "array",
            "title": "Organizations"
          }
        },
        "type": "object",
        "required": [
          "total",
          "organizations"
        ],
        "title": "OrgListResponse"
      },
      "OrgMemberListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/OrgMemberPublic"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "total",
          "members"
        ],
        "title": "OrgMemberListResponse"
      },
      "OrgMemberPublic": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "org_id",
          "role"
        ],
        "title": "OrgMemberPublic",
        "description": "A user's membership in an organization."
      },
      "OrgPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "created_at"
        ],
        "title": "OrgPublic",
        "description": "Public representation of an organization."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  },
...
}